Create video_to_gif.bat
look I didn't want to use the goto make_gif statement but a normal if-else like in datamosh.bat refused to work. I swear it's the truth. Please don't send me links to "Goto Considered Harmful". I never wanted it to be like this. I just wanted to be able to drop a file on the .bat and it then make a GIF.